requirements.txt
# Office AI Agent - full stack dependencies (pinned)
fastapi==0.120.4
uvicorn[standard]==0.38.0
httpx==0.28.1
pydantic==2.12.3
python-dotenv==1.2.1
typing-extensions==4.15.0
langchain==1.0.3
google-generativeai==0.8.5
langchain-google-genai==3.0.0
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
black==25.9.0
isort==7.0.0
flake8==7.3.0
pytest==8.4.2
